Switch access lists to using geo directives
This is the recommended way to handle large access lists in nginx,
and is also an enabling feature to allow IP blocks to use the 444
drop directive for failed IP ranges.

This is a pre-requisite to enabling selective client CA verification
requests.

A side-effect of this is a subtle change in access-list IP ranges:
access-list IP ranges now apply the most specific directive to
determine access.